rename directory, remove dupe focusing, update desc
