+++ packages/react-core/src/demos/CustomMenus/OptionsMenu.md
@@ -7,7 +7,7 @@ source: react-demos
 
 import SearchIcon from '@patternfly/react-icons/dist/esm/icons/search-icon';
 
-As the `OptionsMenu` component is now deprecated, an options menu may now be built using the new suite of `Menu` components. This is showcased below in the example using the new `Dropdpown` component built off of `Menu`.
+As the `OptionsMenu` component is now deprecated, an options menu may now be built using the new suite of `Menu` components. This is showcased below in the example using the new `Select` component built off of `Menu`.
 
 ### Options menu
