+++ packages/react-core/src/demos/CustomMenus/examples/ActionsMenuDemo.tsx
@@ -1,12 +1,5 @@
 import React from 'react';
-import {
-  MenuToggle,
-  MenuItemAction,
-  Dropdown,
-  DropdownGroup,
-  DropdownList,
-  DropdownItem
-} from '@patternfly/react-core';
+import { MenuToggle, MenuItemAction, Select, SelectGroup, SelectList, SelectOption } from '@patternfly/react-core';
 import BarsIcon from '@patternfly/react-icons/dist/esm/icons/bars-icon';
 import ClipboardIcon from '@patternfly/react-icons/dist/esm/icons/clipboard-icon';
 import CodeBranchIcon from '@patternfly/react-icons/dist/esm/icons/code-branch-icon';
@@ -29,19 +22,12 @@ export const ActionsMenuDemo: React.FunctionComponent = () => {
     }
   };
 
-  const onToggleClick = (ev: React.MouseEvent) => {
-    ev.stopPropagation();
-    setTimeout(() => {
-      if (menuRef.current) {
-        const firstElement = menuRef.current.querySelector('li > button:not(:disabled), li > a:not(:disabled)');
-        firstElement && (firstElement as HTMLElement).focus();
-      }
-    }, 0);
+  const onToggleClick = () => {
     setIsOpen(!isOpen);
   };
 
   return (
-    <Dropdown
+    <Select
       isOpen={isOpen}
       ref={menuRef}
       toggle={(toggleRef) => (
@@ -54,9 +40,9 @@ export const ActionsMenuDemo: React.FunctionComponent = () => {
       onSelect={onSelect}
       onOpenChange={(isOpen) => setIsOpen(isOpen)}
     >
-      <DropdownGroup label="Actions">
-        <DropdownList>
-          <DropdownItem
+      <SelectGroup label="Actions">
+        <SelectList>
+          <SelectOption
             isSelected={selectedItems.includes(0)}
             actions={
               <MenuItemAction
@@ -71,8 +57,8 @@ export const ActionsMenuDemo: React.FunctionComponent = () => {
             itemId={0}
           >
             Item 1
-          </DropdownItem>
-          <DropdownItem
+          </SelectOption>
+          <SelectOption
             isDisabled
             isSelected={selectedItems.includes(1)}
             actions={<MenuItemAction icon={<BellIcon aria-hidden />} actionId="alert" aria-label="Alert" />}
@@ -80,24 +66,24 @@ export const ActionsMenuDemo: React.FunctionComponent = () => {
             itemId={1}
           >
             Item 2
-          </DropdownItem>
-          <DropdownItem
+          </SelectOption>
+          <SelectOption
             isSelected={selectedItems.includes(2)}
             actions={<MenuItemAction icon={<ClipboardIcon aria-hidden />} actionId="copy" aria-label="Copy" />}
             itemId={2}
           >
             Item 3
-          </DropdownItem>
-          <DropdownItem
+          </SelectOption>
+          <SelectOption
             isSelected={selectedItems.includes(3)}
             actions={<MenuItemAction icon={<BarsIcon aria-hidden />} actionId="expand" aria-label="Expand" />}
             description="This is a description"
             itemId={3}
           >
             Item 4
-          </DropdownItem>
-        </DropdownList>
-      </DropdownGroup>
-    </Dropdown>
+          </SelectOption>
+        </SelectList>
+      </SelectGroup>
+    </Select>
   );
 };
