+++ packages/react-core/src/demos/CustomMenus/examples/DrilldownMenuDemo.tsx
@@ -27,14 +27,7 @@ export const DrilldownMenuDemo: React.FunctionComponent = () => {
   const toggleRef = React.useRef<HTMLButtonElement>(null);
   const menuRef = React.useRef<HTMLDivElement>(null);
 
-  const onToggleClick = (ev: React.MouseEvent) => {
-    ev.stopPropagation(); // Stop handleClickOutside from handling
-    setTimeout(() => {
-      if (menuRef.current) {
-        const firstElement = menuRef.current.querySelector('li > button:not(:disabled), li > a:not(:disabled)');
-        firstElement && (firstElement as HTMLElement).focus();
-      }
-    }, 0);
+  const onToggleClick = () => {
     setIsOpen(!isOpen);
     setMenuDrilledIn([]);
     setDrilldownPath([]);
