+++ packages/react-core/src/demos/CustomMenus/examples/TreeViewMenuDemo.tsx
@@ -166,14 +166,7 @@ export const TreeViewMenuDemo: React.FunctionComponent = () => {
     );
   };
 
-  const onToggleClick = (ev: React.MouseEvent) => {
-    ev.stopPropagation(); // Stop handleClickOutside from handling
-    setTimeout(() => {
-      if (menuRef.current) {
-        const firstElement = menuRef.current.querySelector('li > button:not(:disabled), li > a:not(:disabled)');
-        firstElement && (firstElement as HTMLElement).focus();
-      }
-    }, 0);
+  const onToggleClick = () => {
     setIsOpen(!isOpen);
   };
